add event listener to body to check if the link is stored uniqueLink array and run loadattr function if not
