+++ demo/index.html
@@ -20,7 +20,7 @@
       </ul>
     </div>
     
-    <script>
+    <script>      
       new PreloadHover({
         defaultDomScope: 
           [
@@ -29,5 +29,6 @@
         linkType: 'both'
         }).start();
     </script>
+    <script src="/src/add-link.js"></script>
   </body>
 </html>
